Merge packages.windows into packages
In doing this we created a windows.treeinfo.json which defines the
windows variants for some packages, and excludes those packages that are
not used on windows (some may going forwards, but a bunch will never be
used).
In merging we also added a couple of windows packages that are excluded
from the linux treeinfo.json files.