Port bootstrap tool to Windows
* The directory locking in bootstrap used a Linux concept that does not
work on Windows. This lock reworked to use portalocker which is a portable
lock that does file locking in a portable way.
* A file is locked within a directory instead of the directory because the
file needed to be opened with the standard python Open function which does
not support direcory opening. Windows does not either.
* The portalocker python library was installed
* A Windows serviced unit file was added that runs some initial initialization
on first use. This is not needed on Linux but the portalocker library does not
work proberly on Windows without this extra initialization