Add Systemd like service manager for Windows
To help support pkgpanda and make the code be shared as much as possible
between Windows and Linux we created a systemctl tool and systemctl-exec
for Windows that supports the necessary subset of functionality that
is needed for pkgpanda to work on Windows for starting and stopping services.
Under the covers systemctl-exec runs as a Windows service and starts and stops
the appropriately configurated dcos executable. systemctl sets up the configuration
for the services and starts and stops the systemctl-exec service.