Windows port of pkgpanda gen
Changes mainly consiste of the following:
* Path changes and script name changes
* switch a few settings off on Windows in calc.py
* Port dcos_install script to Windows for installing a Windows agent
* Package the bootstrap download package into a single compressed tarball including
the docker container and the install script as binary data cannot be added to the end
of a powershell script like that of the Linux equivalent
* enable many tests in gen that work as the results of the work done in this directory and
within pkgpanda itself. Extra changes made to tests include script filenames and the fact that
files cannot be opened twice on Windows, so creation of temporary files are created and then passed
down to the functions after the file is closed. Those files are then deleted when the files are finished
with.