Finish port of pkgpanda to Windows
This change mostly consists of the following:
* Symlinks on Windows need to use Junctions (directory links) and hard links (file links)
* Overload a number of file operations in Python to support the Windows symlinks properly, including
removal and copy, rename files/directories, as well as functions like realpath that needed to be updated for links.
* Usage of temporary files caused problems on Windows as they are opened, then reopened down the callstack
or in a child process resulting in an access denied. They are now created, then passed down, and then deleted
if necessary after they are used.
* Updated to generate the Windows environment files in powershell script.
* Checks have been updated slightly to make .ps1 files Windows specific and .sh files to be linux specific
* File paths in constants.py have been simplified to use abspath() where possible
* Tar files and compression are done via 7-zip on Windows to preserve symlinks properly. Note that 7-zip cannot archive
and compress in one go, so we need to do this in two stages on Windows.
* The build cleaner function on Windows has been taken out of the docker file so we can use the utility functions
necessary to properly delete files in Windows, especially with symlinks. They also perform retries for when the operating
system does not release the files quite quick enough for deletion and renaming.
* Move some docker entrypoint commands back to the builder code itself so tests on Windows can use a generic Windows docker file
and not the actual builder docker file which does not work on CI machines without first doing a build
* Enable all tests that can be enabled for Windows. Most of these changes deal with the following kinds of changes:
  - File names and file path differences
  - Windows variant builds, and linux variants on that platform so Windows variants are not built
  - new tests for the new library functions added (mainly symlinks)